+++ src/Application/Query/Happening/ProgramViewQueryHandler.php
@@ -4,8 +4,13 @@ namespace Proximum\Vimeet\Application\Query\Happening;
 
 use Proximum\Vimeet\Application\Exception\Happening\MissingEventDayConfigurationException;
 use Proximum\Vimeet\Application\View\Happening\ProgramView;
+use Proximum\Vimeet\Domain\Event\GetTimezoneHelper;
+use Proximum\Vimeet\Domain\Model\Sheet;
+use Proximum\Vimeet\Domain\Model\User;
+use Proximum\Vimeet\Domain\Participant\IsParticipantVisio;
 use Proximum\Vimeet\Domain\Repository\Event\DayRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\Unavailability\MassRepositoryInterface;
+use Proximum\Vimeet\Domain\Time\TimeRangeViewTransformer;
 
 class ProgramViewQueryHandler
 {
@@ -24,18 +29,28 @@ class ProgramViewQueryHandler
     /** @var FullHappeningQueryHandler */
     private $fullHappeningQueryHandler;
 
+    /** @var GetTimezoneHelper */
+    private $getTimezoneHelper;
+
+    /** @var IsParticipantVisio */
+    private $isParticipantVisio;
+
     public function __construct(
         DayRepositoryInterface $dayRepository,
         DayViewQueryHandler $dayViewQueryHandler,
         HappeningParticipationQueryHandler $happeningParticipationQueryHandler,
         MassRepositoryInterface $massRepository,
-        FullHappeningQueryHandler $fullHappeningQueryHandler
+        FullHappeningQueryHandler $fullHappeningQueryHandler,
+        GetTimezoneHelper $getTimezoneHelper,
+        IsParticipantVisio $isParticipantVisio
     ) {
         $this->dayRepository = $dayRepository;
         $this->dayViewQueryHandler = $dayViewQueryHandler;
         $this->happeningParticipationQueryHandler = $happeningParticipationQueryHandler;
         $this->massRepository = $massRepository;
         $this->fullHappeningQueryHandler = $fullHappeningQueryHandler;
+        $this->getTimezoneHelper = $getTimezoneHelper;
+        $this->isParticipantVisio = $isParticipantVisio;
     }
 
     /**
@@ -53,6 +68,12 @@ class ProgramViewQueryHandler
             throw new MissingEventDayConfigurationException();
         }
 
+        $timeZone = $this->getTimezoneHelper->getTimezoneByEventAndUser($programViewQuery->event, $programViewQuery->user);
+
+        $timeRangeDays = TimeRangeViewTransformer::fromEventDays($eventDays, $timeZone);
+
+        $translatedTimeZone = $this->getTimezoneHelper->getTimezoneTranslated($timeZone);
+
         $masses = [];
 
         if (null === $programViewQuery->category) {
@@ -63,13 +84,13 @@ class ProgramViewQueryHandler
         }
 
         $dayViews = [];
-        foreach ($eventDays as $day) {
+        foreach ($timeRangeDays as $timeRange) {
             $dayViews[] = $this->dayViewQueryHandler->handle(
                 new DayViewQuery(
                     $programViewQuery->event,
                     $programViewQuery->sheet,
                     $programViewQuery->user,
-                    $day,
+                    $timeRange,
                     $programViewQuery->locale,
                     $programViewQuery->category,
                     $masses
@@ -81,7 +102,12 @@ class ProgramViewQueryHandler
             ? $programViewQuery->category->getTitle($programViewQuery->locale)
             : null;
 
+        $displayTimeZone = $this->displayTimeZone($programViewQuery->sheet, $programViewQuery->user);
+
         $programView = new ProgramView(
+            $displayTimeZone,
+            $translatedTimeZone,
+            $timeZone,
             $dayViews,
             $categoryTitle
         );
@@ -102,5 +128,12 @@ class ProgramViewQueryHandler
         );
 
         return $programView;
+    }
+
+    private function displayTimeZone(Sheet $sheet, User $user): bool
+    {
+        $participant = $sheet->getUserParticipant($user);
+
+        return $participant && $this->isParticipantVisio->isSatisfiedBy($participant) && $participant->getTimezone();
     }
 }