+++ src/Ui/Bundle/EventBundle/Controller/Program/IndexAction.php
@@ -11,6 +11,7 @@ use Proximum\Vimeet\Application\Query\Tip\TipTranslationViewQuery;
 use Proximum\Vimeet\Application\Query\Tip\TipTranslationViewQueryHandler;
 use Proximum\Vimeet\Application\View\Happening\ProgramView;
 use Proximum\Vimeet\Domain\Model\Sheet;
+use Proximum\Vimeet\Domain\Participant\IsParticipantVisio;
 use Proximum\Vimeet\Domain\Participant\ParticipantHelper;
 use Proximum\Vimeet\Ui\Bundle\EventBundle\ParamConverter\EventDomain;
 use Proximum\Vimeet\Ui\Bundle\EventBundle\Security\SheetVoter;
@@ -35,16 +36,21 @@ class IndexAction
     /** @var RouterInterface */
     private $router;
 
+    /** @var IsParticipantVisio */
+    private $isParticipantVisio;
+
     public function __construct(
         AuthorizationCheckerAdapterInterface $authorizationCheckerAdapter,
         QueryBusInterface $queryBus,
         RouterInterface $router,
-        EngineInterface $engine
+        EngineInterface $engine,
+        IsParticipantVisio $isParticipantVisio
     ) {
         $this->authorizationCheckerAdapter = $authorizationCheckerAdapter;
         $this->queryBus = $queryBus;
         $this->engine = $engine;
         $this->router = $router;
+        $this->isParticipantVisio = $isParticipantVisio;
     }
 
     /**
@@ -70,6 +76,21 @@ class IndexAction
 
         $event = $eventDomain->getEvent();
         $user = $userDomain->getUser();
+
+        $participant = $sheet->getUserParticipant($user);
+
+        if ($participant && $this->isParticipantVisio->isSatisfiedBy($participant) && !$participant->getTimezone()) {
+            return new RedirectResponse(
+                $this->router->generate(
+                    'event_participant_timezone',
+                    [
+                        'participant' => $participant->getId(),
+                        'sheet' => $sheet->getId(),
+                        'fromProgram' => '',
+                    ]
+                )
+            );
+        }
 
         try {
             /** @var ProgramView $program */
@@ -104,6 +125,7 @@ class IndexAction
             'program' => $program,
             'isUserAloneParticipant' => $isUserAloneParticipant,
             'tipTranslationViews' => $tipTranslationViews,
+            'participant' => $participant,
         ]));
     }
 }