+++ tests/Application/Query/Happening/DayViewQueryHandlerTest.php
@@ -27,6 +27,7 @@ use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\Type;
 use Proximum\Vimeet\Domain\Model\Unavailability;
 use Proximum\Vimeet\Domain\Repository\HappeningRepositoryInterface;
+use Proximum\Vimeet\Domain\Time\TimeRangeView;
 use Proximum\Vimeet\Tests\Factory\EventFactory;
 use Proximum\Vimeet\Tests\Factory\UserFactory;
 
@@ -40,15 +41,17 @@ class DayViewQueryHandlerTest extends TestCase
         $sheet     = $this->prophesize(Sheet::class);
         $type      = $this->prophesize(Type::class);
         $category  = null;
-        $startTime = new \DateTime('2016-10-12 10:00:00');
-        $endTime   = new \DateTime('2016-10-12 18:00:00');
-        $eventDay  = new Day($event, $startTime, $endTime);
+        $massStartTime = new \DateTime('2016-10-12 11:00:00');
+        $massEndTime   = new \DateTime('2016-10-12 19:00:00');
+        $timeRangeStartTime = new \DateTime('2016-10-12 10:00:00');
+        $timeRangeEndTime   = new \DateTime('2016-10-12 18:00:00');
+        $timeRange = new TimeRangeView($timeRangeStartTime, $timeRangeEndTime);
 
         $sheet->getType()->willReturn($type->reveal());
 
         // Mass
         $categoryMass = new Unavailability\Category($event, 'picto', 'title', 'leftColor', 'rightColor');
-        $mass = new Unavailability\Mass($event, $categoryMass, 'name', $startTime, $endTime, true);
+        $mass = new Unavailability\Mass($event, $categoryMass, 'name', $massStartTime, $massEndTime, true);
 
         // Data
         $beginHappening1 = new \DateTime('2016-10-12 12:00:00');
@@ -112,8 +115,8 @@ class DayViewQueryHandlerTest extends TestCase
 
         $massView = new MassUnavailabilityView(
             1,
-            $startTime,
-            $endTime,
+            $massStartTime,
+            $massEndTime,
             'title',
             'description',
             'picto',
@@ -124,8 +127,8 @@ class DayViewQueryHandlerTest extends TestCase
         );
 
         $expected = new DayView(
-            $startTime,
-            $endTime,
+            $timeRangeStartTime,
+            $timeRangeEndTime,
             $event->getConfiguration()->getScheduleScale(),
             [
                 $happeningView1,
@@ -143,7 +146,7 @@ class DayViewQueryHandlerTest extends TestCase
         $happeningRepository->findByEventAndTypeAndDayAndCategory(
             $event,
             $type->reveal(),
-            $eventDay->getDay(),
+            $timeRange->getBegin(),
             $category
         )->shouldBeCalled()->willReturn($happenings);
 
@@ -180,7 +183,7 @@ class DayViewQueryHandlerTest extends TestCase
             $event,
             $sheet->reveal(),
             $user,
-            $eventDay,
+            $timeRange,
             'fr',
             $category,
             [$mass]